test: split codex credit refresh coverage
